requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.4.2
blinker==1.8.2
braceexpand==0.1.7
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
contourpy==1.2.1
cycler==0.12.1
discord.py @ git+https://github.com/Rapptz/discord.py.git@b0024dc86652db19cad16296d6c16c6272b2eaa3
dnspython==2.6.1
docopt==0.6.2
exceptiongroup==1.2.1
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
import_expression==2.0.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jishaku @ git+https://github.com/Gorialis/jishaku.git@f68f4395cc58fca247b86967b52b35d95f2b7cd3
kiwisolver==1.4.5
line_profiler==4.1.3
MarkupSafe==2.1.5
matplotlib==3.9.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
pip-upgrader==1.4.15
platformdirs==4.2.2
priority==2.0.0
prometheus_client==0.20.0
psutil==5.9.8
pymongo==4.7.3
pyparsing==3.1.2
pypxl==0.2.3
python-dateutil==2.9.0
pyzmq==26.0.3
Quart==0.19.6
quart-cors==0.7.0
requests==2.32.3
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.9.0
taskgroup==0.0.0a4
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.1
urllib3==2.2.1
uvicorn==0.30.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
